feat: 主动安全防护系统附件服务器

+++ active-safety/src/main/java/com/avenger/jt808/handler/AttachmentInformationHandler.java
@@ -14,7 +14,7 @@ import reactor.core.publisher.Mono;
  */
 public class AttachmentInformationHandler implements MessageHandler {
     @Override
-    public int getId() {
+    public short getId() {
         return 0x1211;
     }
